Updated RTC
removed dependencies on other files

+++ RTC.h
@@ -7,8 +7,8 @@ extern Calendar GetCurrentCalendar();
 extern int GetCurrentHour();
 extern int GetCurrentMin();
 extern int GetCurrentSec();
-extern int BCDtoInt( int bcd );
-extern int IntToBCD( int val );
+extern uint8_t BCDtoInt( uint8_t bcd );
+extern uint8_t IntToBCD( uint8_t val );
 extern void SetCalendar( Calendar calendar );
 
 #endif
